+++ lib/open_exchange_rates/updater.ex
@@ -4,43 +4,50 @@ defmodule OpenExchangeRates.Updater do
   """
 
   require Logger
-  use GenServer
 
   @update_interval_in_seconds 60
 
-  @doc false
-  def start_link do
-    {:ok, pid} = GenServer.start_link(__MODULE__, nil, name: __MODULE__)
-    maybe_load_data_from_disk
-    check_for_update
-    {:ok, pid}
-  end
+  def cache_file, do: Application.get_env(:open_exchange_rates, :cache_file) || (File.cwd! <> "/priv/latest.json")
 
   @doc false
-  def init(_opts) do
-    :timer.apply_interval(:timer.seconds(@update_interval_in_seconds), __MODULE__, :check_for_update, [])
-    {:ok, %{last_updated_at: 0}}
+  def start do
+    load_data_from_disk
+    check_for_update
+    :ok
   end
 
   @doc false
   def check_for_update do
-    last_updated_at = GenServer.call(__MODULE__, {:last_updated_at})
-    diff =  (:os.system_time(:seconds) - last_updated_at) / 60
-    cache_time = Application.get_env(:open_exchange_rates, :cache_time_in_minutes, 1440)
-    if diff >= cache_time, do: update
+    cache_time = Application.get_env(:open_exchange_rates, :cache_time_in_minutes, 1440) * 60
+
+    if OpenExchangeRates.Cache.cache_age >= cache_time, do: update
+    Process.sleep(@update_interval_in_seconds * 1000)
+    check_for_update
   end
 
-  defp maybe_load_data_from_disk do
-    case File.exists?(OpenExchangeRates.Cache.file) do
-      true -> OpenExchangeRates.Cache.file |> File.read! |> Poison.decode! |> update_cache!
-      false -> update
+  def load_data_from_disk do
+    if File.exists?(cache_file()) do
+      cached_at = File.lstat!(cache_file(), [time: :posix]).mtime
+      cache_file()
+      |> File.read!
+      |> Poison.decode!
+      |> update_cache!(cached_at)
+      :ok
+    else
+      {:error, "could not find #{cache_file()} which you defined in your config"}
+    end
+  end
+
+  def load_data_from_disk! do
+    case load_data_from_disk do
+      :ok -> :ok
+      {:error, message} -> raise(message)
     end
   end
 
   defp write_to_disk!(data) do
-    GenServer.call(__MODULE__, {:set_last_updated_at, :os.system_time(:seconds)})
     json = Poison.encode!(data)
-    File.write!(OpenExchangeRates.Cache.file, json)
+    File.write!(cache_file(), json)
     data
   end
 
@@ -52,17 +59,11 @@ defmodule OpenExchangeRates.Updater do
     end
   end
 
-  defp update_cache!(%{"rates" => rates}), do: GenServer.call(OpenExchangeRates.Cache, {:update!, rates})
-  defp update_cache!(data), do: raise "Data was corrupted ? #{inspect data}"
-
-  def handle_call({:last_updated_at}, _caller, state) do
-    timestamp = Map.get(state, :last_updated_at, 0)
-    {:reply, timestamp, state}
+  defp update_cache!(data, cached_at \\ nil)
+  defp update_cache!(%{"rates" => rates}, cached_at) do
+    cached_at = cached_at || :os.system_time(:seconds)
+    GenServer.call(OpenExchangeRates.Cache, {:update!, rates, cached_at})
   end
-
-  def handle_call({:set_last_updated_at, timestamp}, _caller, state) do
-    state = Map.put(state, :last_updated_at, timestamp)
-    {:reply, :ok, state}
-  end
+  defp update_cache!(data, _cached_at), do: raise "Data was corrupted ? #{inspect data}"
 
 end