+++ test/open_exchange_rates_test.exs
@@ -39,9 +39,43 @@ defmodule OpenExchangeRatesTest do
     assert  {:ok, 118} = OpenExchangeRates.convert_cents(100, "GBP", "EUR")
     assert  {:ok, 118} = OpenExchangeRates.convert_cents(100, "gbp", "eur")
     assert  {:ok, 118} = OpenExchangeRates.convert_cents(100, :gbp, :eur)
+    assert  {:ok, 1.1790803434757773} = OpenExchangeRates.convert(1, "GBP", "EUR")
     assert  {:ok, 1.1790803434757773} = OpenExchangeRates.convert(1.0, "GBP", "EUR")
     assert  {:ok, 1.1790803434757773} = OpenExchangeRates.convert(1.0, "gbp", "eur")
     assert  {:ok, 1.1790803434757773} = OpenExchangeRates.convert(1.0, :gbp, :eur)
+    assert  {:ok, 1.1790803434757773} = OpenExchangeRates.conversion_rate(:gbp, :eur)
+    assert  {:ok, 1.1790803434757773} = OpenExchangeRates.conversion_rate("gbp", "eur")
+    assert  {:ok, 1.1790803434757773} = OpenExchangeRates.conversion_rate("GBP", "EUR")
   end
 
+  test "it should return the cache age" do
+    assert is_integer(OpenExchangeRates.cache_age)
+  end
+
+  test "it should return an error when formatting an unknown currency" do
+    assert {:error, "unknown currency: BLA"} = OpenExchangeRates.convert_and_format(123456789, :EUR, :BLA)
+  end
+  test "it should convert and return a formatted string" do
+    assert {:ok, "£104,706,002.17"} = OpenExchangeRates.convert_and_format(123456789, :EUR, :GBP)
+    assert "£104,706,002.17" = OpenExchangeRates.convert_and_format!(123456789, :EUR, :GBP)
+
+    assert {:ok, "£1,047,060.02"} = OpenExchangeRates.convert_cents_and_format(123456789, :EUR, :GBP)
+    assert "£1,047,060.02" = OpenExchangeRates.convert_cents_and_format!(123456789, :EUR, :GBP)
+  end
+
+  test "it show raise when convert! with unknown currency" do
+    assert_raise RuntimeError, "unknown currency: BLA", fn -> OpenExchangeRates.convert!(100, :EUR, :BLA) end
+  end
+
+  test "it show raise when convert_cents! with unknown currency" do
+    assert_raise RuntimeError, "unknown currency: BLA", fn -> OpenExchangeRates.convert_cents!(100, :EUR, :BLA) end
+  end
+
+  test "it show raise when convert_and_format! with unknown currency" do
+    assert_raise RuntimeError, "unknown currency: BLA", fn -> OpenExchangeRates.convert_and_format!(100, :EUR, :BLA) end
+  end
+
+  test "it show raise when convert_cents_and_format! with unknown currency" do
+    assert_raise RuntimeError, "unknown currency: BLA", fn -> OpenExchangeRates.convert_cents_and_format!(100, :EUR, :BLA) end
+  end
 end